reordered restaurants where they are displayed
